+++ wostools/__init__.py
@@ -2,10 +2,10 @@
 
 __author__ = """Core of Science"""
 __email__ = "dev@coreofscience.com"
-__version__ = "2.0.7"
+__version__ = "3.0.0"
 
 from wostools.article import Article
-from wostools.lazy import LazyCollection
 from wostools.cached import CachedCollection
+from wostools.cached import CachedCollection as Collection
 
-__all__ = ["CachedCollection", "LazyCollection", "Article"]
+__all__ = ["CachedCollection", "Collection", "Article"]